Ran final experiments

+++ experiment.py
@@ -4,7 +4,7 @@
 # Created:
 #   24/06/2021, 16:15:59
 # Last edited:
-#   25/06/2021, 01:58:17
+#   25/06/2021, 09:31:29
 # Auto updated?
 #   Yes
 #
@@ -999,7 +999,7 @@ def amazing_graph(density_list):
 
 ### EXPERIMENTS
 
-def run_and_save_experiment(output_dir, alpha, threshold, nut_inflow, N = 100, eta = 1, omega = 0.7, iterations = 100, test_iterations = 10):
+def run_and_save_experiment(output_dir, alpha, threshold, nut_inflow, N = 100, eta = 1, omega = 0.7, iterations = 100, test_iterations = 10, compress = True):
     """
         Function that runs the whole simulation multiple times, and saves all
         results to the given directory.
@@ -1056,12 +1056,13 @@ def run_and_save_experiment(output_dir, alpha, threshold, nut_inflow, N = 100, e
         prgs.update()
     
     # When done, compress the whole folder
-    print(f"Compressing resuts to '{output_dir}.zip'...")
-    os.system(f"zip -r9 {output_dir}.zip {output_dir}")
-    # Delete the old folder
-    os.system(f"rm -rf {output_dir}")
+    if compress:
+        print(f"Compressing resuts to '{output_dir}.zip'...")
+        os.system(f"zip -r9 {output_dir}.zip {output_dir}")
+        # Delete the old folder
+        os.system(f"rm -rf {output_dir}")
 
-def vary_param(output_dir, values, test_iterations = 10):
+def vary_param(output_dir, values, test_iterations = 10, compress = True):
     """
         Function that tests the given parameter for the given values.
 
@@ -1134,7 +1135,7 @@ def vary_param(output_dir, values, test_iterations = 10):
 
         # Launch the experiment
         print(f"Running for alpha={alpha_value}, threshold={threshold_value}, nut_inflow={nut_inflow_value}:")
-        run_and_save_experiment(paths[i], alpha_value, threshold_value, nut_inflow_value, test_iterations=test_iterations)
+        run_and_save_experiment(paths[i], alpha_value, threshold_value, nut_inflow_value, test_iterations=test_iterations, compress=compress)
 
 
 
@@ -1147,10 +1148,10 @@ if __name__ == "__main__":
         # "alpha": [1.0],
         "alpha": 0.0,
         # "threshold": [0.0, 0.05, 0.0501, 0.0502, 0.0503, 0.0504, 0.0505, 0.0506, 0.0507, 0.0508, 0.0509, 0.051, 0.5],
-        # "threshold": [0.0, 0.0503, 0.0504, 0.0505, 0.0506, 0.0507, 0.1],
-        "threshold": [0.01],
+        # "threshold": [0.0, 0.0503, 0.0504, 0.0505, 0.0506, 0.0507, 0.01],
+        "threshold": [0.1],
         "nut_inflow": 0.3
     }
 
     # Run the alphas
-    vary_param("/run/media/lut_99/Data/Temp/threshold", configs, test_iterations=10)
+    vary_param("/run/media/lut_99/Data/Temp/threshold", configs, test_iterations=10, compress = False)
